# app/static/swagger.yaml
...
swagger: '2.0'
info:
  title: "Service Ticket Management System API"
  description: "API to manage a Mechanic Service Ticket System"
  version: "1.0.0"
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"

# =========================
# Security Definitions
# =========================
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header

# =========================
# Paths
# =========================
paths: 
...
  /service-tickets: #Service Tickets CRUD
    post: #Create
      tags:
        - Service Tickets
      summary: Create a new service ticket
      description: This endpoint allows you to create a new service ticket.
      parameters:
        - in: body
          name: body
          description: Service ticket object that needs to be added
          required: true
          schema:
            $ref: '#/definitions/ServiceTicketPayload'
      responses:
        201:
          description: Service ticket created successfully
          schema:
            $ref: '#/definitions/ServiceTicketResponse'
          examples:
            application/json:
              {
                "id": 2,
                "customer_id": 1,
                "customer": {
                  "email": "jd@app.com",
                  "id": 1,
                  "name": "John Doe",
                  "phone": "123-456-789"
                },
                "mechanics": [
                    {
                        "email": "mike@app.com",
                        "id": 1,
                        "name": "Mike Store",
                        "phone": "123-456-789",
                        "salary": "50000.00"
                    }
                ],
                "service_date": "2025-03-20",
                "service_desc": "This is a description",
                "vin": "CMD12456",
                "ticket_items": []
              }
        400:
          description: Invalid customer ID or mechanic ID
          schema:
            $ref: '#/definitions/MessageSchemaResponse'
          examples:
            application/json:
              invalid_customer_id:
                message: Invalid customer ID.
                status: error
              invalid_mechanic_id:
                message: Invalid mechanic ID.
                status: error
    get: #Read
      tags:
        - Service Tickets
      summary: Get service tickets (with pagination)
      description: This endpoint returns a list of service tickets with pagination. 
        **You can use the `page` and `per_page` query parameters to control pagination.**
      parameters:
        - in: query
          name: page
          type: integer
          description: Page number for pagination.
        - in: query
          name: per_page
          type: integer
          description: Number of service tickets per page.
      responses:
        200:
          description: Return all service tickets
          schema:
            $ref: '#/definitions/ServiceTicketsListResponse'
          examples:
            # application/json:
            #   [
            #     {
            #       "customer": {
            #         "email": "jd@app.com",
            #         "id": 1,
            #         "name": "John Doe",
            #         "phone": "123-456-789"
            #       },
            #       "id": 1,
            #       "mechanics": [
            #           {
            #               "email": "mike@app.com",
            #               "id": 1,
            #               "name": "Mike Store",
            #               "phone": "123-456-789",
            #               "salary": "50000.00"
            #           }
            #       ],
            #       "service_date": "2025-03-20",
            #       "service_desc": "This is a description",
            #       "ticket_items": [],
            #       "vin": "CMD12456"
            #     }
            #   ]
        
  #---------- Service Tickets Add Single Mechanic ---------
  /service-tickets/{ticket_id}/add-mechanic/{mechanic_id}:
    put: #Update
      tags:
        - Service Tickets
      summary: Add a mechanic to a service ticket
      description: This endpoint allows you to add a mechanic to an existing service ticket.
      parameters:
        - in: path
          name: ticket_id
          required: true
          schema:
            type: integer
            minimum: 1
          description: The ID of the service ticket to update.
        - in: path
          name: mechanic_id
          required: true
          schema:
            type: integer
            minimum: 1
          description: The ID of the mechanic to add.
      responses:
        200:
          description: Mechanic added successfully
          schema:
            $ref: '#/definitions/ServiceTicketUpdateResponse'
          examples:
            application/json:
              {
                "message": "Mechanic added successfully.",
                "status": "success",
                "service_ticket":{
                  "customer": {
                    "email": "jd@app.com",
                    "id": 1,
                    "name": "John Doe",
                    "phone": "123-456-789"
                  },
                  "id": 2,
                  "mechanics": [
                      {
                          "email": "mike@app.com",
                          "id": 1,
                          "name": "Mike Store",
                          "phone": "123-456-789",
                          "salary": "50000.00"
                      }
                  ],
                  "service_date": "2025-03-20",
                  "service_desc": "This is a description",
                  "ticket_items": [],
                  "vin": "CMD12456"
                }
              }
        400:
          description: Mechanic already assigned to the ticket
          schema:
            $ref: '#/definitions/MessageSchemaResponse'
          examples:
            application/json:
              message: Mechanic already assigned to the ticket.,
              status": error
        404:
          description: Service ticket or mechanic not found
          schema:
            $ref: '#/definitions/MessageSchemaResponse'
          examples:
            application/json:
              message: Service ticket or mechanic not found.,
              status": error

  #---------- Service Tickets Remove Single Mechanic ---------            
  /service-tickets/{ticket_id}/remove-mechanic/{mechanic_id}:
    put: #Update
      tags:
        - Service Tickets
      summary: Remove a mechanic from a service ticket
      description: This endpoint allows you to remove a mechanic from an existing service ticket.
      parameters:
        - in: path
          name: ticket_id
          required: true
          schema:
            type: integer
            minimum: 1
          description: The ID of the service ticket to update.
        - in: path
          name: mechanic_id
          required: true
          schema:
            type: integer
            minimum: 1
          description: The ID of the mechanic to remove.
      responses:
        200:
          description: Mechanic removed successfully
          schema:
            $ref: '#/definitions/ServiceTicketUpdateResponse'
          examples:
            application/json:
              {
                "message": "Mechanic removed successfully.",
                "status": "success",
                "service_ticket":{
                  "customer": {
                    "email": "jd@app.com",
                    "id": 1,
                    "name": "John Doe",
                    "phone": "123-456-789"
                  },
                  "id": 2,
                  "mechanics": [
                      {
                          "email": "mike@app.com",
                          "id": 1,
                          "name": "Mike Store",
                          "phone": "123-456-789",
                          "salary": "50000.00"
                      }
                  ],
                  "service_date": "2025-03-20",
                  "service_desc": "This is a description",
                  "ticket_items": [],
                  "vin": "CMD12456"
                }
              }
        400:
          description: Mechanic not included on this ticket
          schema:
            $ref: '#/definitions/MessageSchemaResponse'
          examples:
            application/json:
              message: Mechanic not included on this ticket.
              status: error   
        404:
          description: Service ticket or mechanic not found
          schema:
            $ref: '#/definitions/MessageSchemaResponse'
          examples:
            application/json:
              message: Service ticket or mechanic not found.
              status: error

  #---------- Service Tickets Add Serialized Part ID ---------             
  /service-tickets/{ticket_id}/add-part/{part_id}:
    put:
      tags:
        - Service Tickets
      summary: Add a part to a service ticket
      description: This endpoint allows you to add a part to an existing service ticket.
      parameters:
        - in: path
          name: ticket_id
          required: true
          schema:
            type: integer
            minimum: 1
          description: The ID of the service ticket to update.
        - in: path
          name: part_id
          required: true
          schema:
            type: integer
            minimum: 1
          description: The ID of the part to add.
      responses:
        200:
          description: Serialized serialized part added successfully
          schema:
            $ref: '#/definitions/ServiceTicketUpdateResponse'
          examples:
            application/json:
              {
                "message": "Serialized part 1 successfully added to ticket.",
                "status": "success",
                "service_ticket":{
                  "customer": {
                    "email": "jd@app.com",
                    "id": 1,
                    "name": "John Doe",
                    "phone": "123-456-789"
                  },
                  "id": 2,
                  "mechanics": [
                      {
                          "email": "mike@app.com",
                          "id": 1,
                          "name": "Mike Store",
                          "phone": "123-456-789",
                          "salary": "50000.00"
                      }
                  ],
                  "service_date": "2025-03-20",
                  "service_desc": "This is a description",
                  "ticket_items": [{
                      "desc_id": 1,
                      "ticket_id": 2,
                      "description": {
                          "id": 1,
                          "name": "Brake Pad",
                          "brand": "Brand A",
                          "price": 100.00
                      }
                  }],
                  "vin": "CMD12456"
                }
              }
        400:
          description: Serialized part already assigned to a ticket
          schema:
            $ref: '#/definitions/MessageSchemaResponse'
          examples:
            application/json:
              message: Serialized part already assigned to a ticket
              status: error
        404:
          description: Service ticket or serialized part not found
          schema:
            $ref: '#/definitions/MessageSchemaResponse'
          examples:
            application/json:
              message: Service ticket or serialized part not found.
              status: error

  #---------- Service Tickets Remove Serialized Part ID ---------
  /service-tickets/{ticket_id}/remove-part/{part_id}:
    put: #Update
      tags:
        - Service Tickets
      summary: Remove a part from a service ticket
      description: This endpoint allows you to remove a part from an existing service ticket.
      parameters:
        - in: path
          name: ticket_id
          required: true
          schema:
            type: integer
            minimum: 1
          description: The ID of the service ticket to update.
        - in: path
          name: part_id
          required: true
          schema:
            type: integer
            minimum: 1
          description: The ID of the part to remove.
      responses:
        200:
          description: Serialized part removed successfully
          schema:
            $ref: '#/definitions/ServiceTicketUpdateResponse'
          examples:
            application/json:
              {
                "message": "Part 1 successfully removed to ticket.",
                "status": "success",
                "service_ticket":{
                  "customer": {
                    "email": "jd@app.com",
                    "id": 1,
                    "name": "John Doe",
                    "phone": "123-456-789"
                  },
                  "id": 2,
                  "mechanics": [
                      {
                          "email": "mike@app.com",
                          "id": 1,
                          "name": "Mike Store",
                          "phone": "123-456-789",
                          "salary": "50000.00"
                      }
                  ],
                  "service_date": "2025-03-20",
                  "service_desc": "This is a description",
                  "ticket_items": [],
                  "vin": "CMD12456"
                }
              }
        400:
          description: Serialized part not included on this ticket
          schema:
            $ref: '#/definitions/MessageSchemaResponse'
          examples:
            application/json:
              message: Serialized part not included on this ticket.
              status: error
        404:
          description: Service ticket or part not found
          schema:
            $ref: '#/definitions/MessageSchemaResponse'
          examples:
            application/json:
              message: Service ticket or part not found.
              status: error
              
  #---------- Service Tickets Add/Remove Multiple Mechanics ---------
  /service-tickets/{id}/edit-mechanics:
    put: #Update
      tags:
        - Service Tickets
      summary: Edit service ticket by ID
      description: Edit a service ticket's details by its unique ID.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            minimum: 1
          description: The ID of the item to update.
        - in: body
          name: body
          description: Information used to update a service ticket
          required: true
          schema:
            type: object
            properties:
              add_mechanic_ids:
                type: array
                items:
                  type: integer
              remove_mechanic_ids:
                type: array
                items:
                  type: integer
      responses:
        200:
          description: Update service ticket successfully
          schema:
            $ref: '#/definitions/ServiceTicketUpdateResponse'
          examples:
            application/json:
              # {
              #   "message": "Part 1 successfully removed to ticket.",
              #   "status": "success",
              #   "service_ticket":{
              #     "customer": {
              #       "email": "jd@app.com",
              #       "id": 1,
              #       "name": "John Doe",
              #       "phone": "123-456-789"
              #     },
              #     "id": 2,
              #     "mechanics": [
              #         {
              #             "email": "mike@app.com",
              #             "id": 1,
              #             "name": "Mike Store",
              #             "phone": "123-456-789",
              #             "salary": "50000.00"
              #         }
              #     ],
              #     "service_date": "2025-03-20",
              #     "service_desc": "This is a description",
              #     "ticket_items": [],
              #     "vin": "CMD12456"
              #   }
              # }
        400:
          description: Mechanic assigned/unassigned in this ticket or mechanic does not exist
          schema:
            $ref: '#/definitions/MessageSchemaResponse'
          examples:
            application/json:
              you_add_the_assigned_mechanic:
                  message: "Mechanic already assigned to this ticket."
                  status: "error"
              you_remove_the_unassigned_mechanic:
                  message: "Mechanic not assigned to this ticket."
                  status: "error"
              mechanic_does_not_exist:
                  message: "Mechanic does not exist."
                  status: "error"
        404:
          description: Service ticket not found
          schema:
            $ref: '#/definitions/MessageSchemaResponse'
          examples:
            application/json:
                message: Service ticket not found.
                status: error

  #---------- Service Tickets Add Parts to Cart ---------       
  /service-tickets/{ticket_id}/add-to-cart/{part_id}:
    put: #Update
      tags:
        - Service Tickets
      summary: Add parts to a service ticket
      description: This endpoint allows you to add parts to an existing service ticket.
      parameters:
        - in: path
          name: ticket_id
          required: true
          schema:
            type: integer
            minimum: 1
          description: The ID of the service ticket to update.
        - in: path
          name: part_id
          required: true
          schema:
            type: integer
            minimum: 1
          description: The ID of the part to add.
        - in: body
          name: body
          description: Quantity of parts to add (default is 1)
          required: false
          schema:
            type: object
            properties:
              quantity:
                type: integer
                minimum: 1
                default: 1
      responses:
        200:
          description: Serialized parts added successfully
          schema:
            $ref: '#/definitions/ServiceTicketUpdateResponse'
          examples:
            application/json:
        400:
          description: Quantity is more than the available stock
          schema:
            $ref: '#/definitions/MessageSchemaResponse'
          examples:
            application/json:
              message: Only 5 stock(s) available for this part.
              status: error
        404:
          description: Invalid ticket_id or part_id
          schema:
            $ref: '#/definitions/MessageSchemaResponse'
          examples:
            application/json:
              message: Invalid ticket_id or part_id.
              status: error

  #---------- Service Tickets Delete ---------         
  /service-tickets/{id}:
    delete: #Delete
      tags:
        - Service Tickets
      summary: Delete a service ticket
      description: This endpoint allows you to delete a service ticket.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            minimum: 1
          description: The ID of the item to retrieve.
      responses:
        200:
          description: Service ticket deleted successfully
          schema:
            $ref: '#/definitions/MessageSchemaResponse'
          examples:
            application/json:
              status: success
              message: Successfully deleted service ticket 1  
        404:
          description: Service ticket not found
          schema:
            $ref: '#/definitions/MessageSchemaResponse'
          examples:
            application/json:
              message: Service ticket not found.
              status: error

             
# =========================
# Definitions (Schemas)
# =========================          
definitions: 
...
  MessageSchemaResponse:
    type: object
    properties:
      message:
        type: string
      status:
        type: string

  # --------- Mechanic Schemas ---------
...
  MechanicResponse:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      email:
        type: string
      salary:
        type: integer
      service_tickets:
        type: array     
...
  CustomerResponse:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      phone:
        type: string
      email:
        type: string
...
  PartDescriptionResponse:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      brand:
        type: string
      price:
        type: integer
...
  SerializedPartResponse:
    type: object
    properties:
      id:
        type: integer
      desc_id:
        type: integer
      ticket_id:
        type: integer
      description:
        $ref: '#/definitions/PartDescriptionResponse'
...
  ServiceTicketPayload:
    type: object
    properties:
      customer_id:
        type: integer
      mechanic_ids:
        type: array
        items:
          type: integer
      service_date:
        type: string
      service_desc:
        type: string
      vin:
        type: string
  ServiceTicketResponse:
    type: object
    properties:
      id:
        type: integer
      customer_id:
        type: integer
      customer:
       $ref: '#/definitions/CustomerResponse'
      mechanics:
        type: array
        items:
          $ref: '#/definitions/MechanicResponse'
      ticket_items:
        type: array
        items:
          $ref: '#/definitions/SerializedPartResponse'
      service_date:
        type: string
      service_desc:
        type: string
      vin:
        type: string
  ServiceTicketsListResponse:
    type: array
    items:
      $ref: '#/definitions/ServiceTicketResponse'
  ServiceTicketUpdateResponse:
    type: object
    properties:
      message:
        type: string
      status:
        type: string
      service_ticket:
        $ref: '#/definitions/ServiceTicketResponse'